+++ src/index.mts
@@ -1,6 +1,8 @@
 #!/usr/bin/env node
 import meow from 'meow';
-import { convert } from './quiver-markdown.mjs';
+
+import { convert } from './quiver-to-obsidian-exporter.mjs';
+
 
 const cli = meow(`
 	Usage
